progress on separating query documents from aggregation expressions
